Break each rule's tests into their own file!
