+++ packages/perseus-linter/src/rules/absolute-url.test.ts
@@ -1,0 +1,35 @@
+import {expectWarning, expectPass} from "../__tests__/test-utils";
+
+import absoluteUrlRule from "./absolute-url";
+
+describe("absolute-url", () => {
+    expectWarning(absoluteUrlRule, [
+        // Warn about absolute khanacademy.org urls
+        "[target](http://khanacademy.org/about)",
+        "[target](https://khanacademy.org/about)",
+        "[target](http://www.khanacademy.org/about)",
+        "[target](https://www.khanacademy.org/about)",
+        "[target](http://es.khanacademy.org/about)",
+        "[target](https://es.khanacademy.org/about)",
+        "[target](//www.khanacademy.org/about)",
+        "[target](//www.khanacademy.org/about)",
+
+        // We should get the same warnings for images
+        "![alt text](http://khanacademy.org/about)",
+        "![alt text](https://www.khanacademy.org/about)",
+        "![alt text](https://es.khanacademy.org/about)",
+    ]);
+    expectPass(absoluteUrlRule, [
+        "[target](/about)", // relative URLs okay
+        "[target](https://kasandbox.org/path)",
+        "[target](https://fastly.kastatic.org/path)",
+        "[target](https://cdn.kastatic.org/path)",
+        "[target](https://ka-perseus-images.s3.amazonaws.com/path)",
+        "[target](https://ka-youtube-converted.storage.googleapis.com)",
+
+        // Same warnings for images
+        "![alt text](/about)",
+        "![alt text](https://cdn.kastatic.org/path)",
+        "![alt text](https://ka-perseus-images.s3.amazonaws.com/path)",
+    ]);
+});
