+++ packages/perseus-linter/src/rules/heading-level-1.test.ts
@@ -1,0 +1,8 @@
+import {expectWarning, expectPass} from "../__tests__/test-utils";
+
+import headingLevel1Rule from "./heading-level-1";
+
+describe("heading-level-1", () => {
+    expectWarning(headingLevel1Rule, "# Level 1 heading");
+    expectPass(headingLevel1Rule, "## Level 1 heading\n\n### Level 3 heading");
+});
