+++ packages/perseus-linter/src/rules/math-starts-with-space.test.ts
@@ -1,0 +1,30 @@
+import {expectWarning, expectPass} from "../__tests__/test-utils";
+
+import mathStartsWithSpaceRule from "./math-starts-with-space";
+
+describe("math-starts-with-space", () => {
+    expectWarning(mathStartsWithSpaceRule, [
+        "foo$~ x$bar",
+        "$\\qquad x$",
+        "$\\quad x$",
+        "$\\, x$",
+        "$\\; x$",
+        "$\\: x$",
+        "$\\ x$",
+        "$\\! x$",
+        "$\\enspace x$",
+        "$\\phantom{xyz} x$",
+    ]);
+    expectPass(mathStartsWithSpaceRule, [
+        "$a~ x$",
+        "$a\\qquad x$",
+        "$a\\quad x$",
+        "$a\\, x$",
+        "$a\\; x$",
+        "$a\\: x$",
+        "$a\\ x$",
+        "$a\\! x$",
+        "$a\\enspace x$",
+        "$a\\phantom{xyz} x$",
+    ]);
+});
